Remove the link to Binder.org and Binder Badge
I tried using Binder to allow dynamic execution of this notebook. However, it failed to create a usable docker container, or perhaps it was just too big to execute on the freely available Binder resources. For whatever reasons , it didn't work as well as I had hoped it would. In fact, I tried a small toy notebook, which i did get to work, and indeed was able to dynamically add some math equations, and to do some plotting. So, the things I'm doing in this notebook should have worked as well. Even the little, toy notebook ran quite slowly...this notebook will have to wait for a faster Binder, or until I figure out how to run it more efficiently.