Who ate my Figures?
Having trouble making my figures render in markdown cells here. It works on my local machine.
That is sort of a problem as the whole idea was to be able to share the math and figures with the code.

Odds are that it is my fault. There is probably a better way to do it, and a way that will work here as 
well as locally.